обновление интерфейса, добавлен offline HC

+++ fonts-load.sh
@@ -2,7 +2,10 @@
 
 DIR=./dist/public
 
-mkdir -p $DIR
+if [ ! -f "$DIR/UcC73FwrK3iLTeHuS_nVMrMxCp50SjIa0ZL7SUc.woff2" ]; then
 
-wget -O $DIR/UcC73FwrK3iLTeHuS_nVMrMxCp50SjIa0ZL7SUc.woff2 "https://fonts.gstatic.com/s/inter/v18/UcC73FwrK3iLTeHuS_nVMrMxCp50SjIa0ZL7SUc.woff2"
-wget -O $DIR/UcC73FwrK3iLTeHuS_nVMrMxCp50SjIa1ZL7.woff2 "https://fonts.gstatic.com/s/inter/v18/UcC73FwrK3iLTeHuS_nVMrMxCp50SjIa1ZL7.woff2"
+    mkdir -p $DIR
+
+    wget -O $DIR/UcC73FwrK3iLTeHuS_nVMrMxCp50SjIa0ZL7SUc.woff2 "https://fonts.gstatic.com/s/inter/v18/UcC73FwrK3iLTeHuS_nVMrMxCp50SjIa0ZL7SUc.woff2"
+    wget -O $DIR/UcC73FwrK3iLTeHuS_nVMrMxCp50SjIa1ZL7.woff2 "https://fonts.gstatic.com/s/inter/v18/UcC73FwrK3iLTeHuS_nVMrMxCp50SjIa1ZL7.woff2"
+fi
